backend/requirements.txt
fastapi[standard]==0.115.0
websockets==12.0
httpx==0.27.2
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.9.2
